# server/docs/swagger.yaml
definitions:
...
  EngineeringEducationDatabase.SocialServiceDatabase:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      industryInnovationCenterCount:
        description: 行业协同创新中心数或产学研基地数
        type: integer
      regionalInnovationCenterCount:
        description: 区域协同创新中心数或产学研基地数
        type: integer
      universityName:
        description: 学校名称
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
    required:
    - industryInnovationCenterCount
    - regionalInnovationCenterCount
    - universityName
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.6.1
paths:
...
  /SSD/createSocialServiceDatabase:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建社会服务库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.SocialServiceDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建社会服务库
      tags:
      - SocialServiceDatabase
  /SSD/deleteSocialServiceDatabase:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除社会服务库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.SocialServiceDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除社会服务库
      tags:
      - SocialServiceDatabase
  /SSD/deleteSocialServiceDatabaseByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除社会服务库
      tags:
      - SocialServiceDatabase
  /SSD/findSocialServiceDatabase:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - description: 行业协同创新中心数或产学研基地数
        in: query
        name: industryInnovationCenterCount
        required: true
        type: integer
      - description: 区域协同创新中心数或产学研基地数
        in: query
        name: regionalInnovationCenterCount
        required: true
        type: integer
      - description: 学校名称
        in: query
        name: universityName
        required: true
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询社会服务库
      tags:
      - SocialServiceDatabase
  /SSD/getSocialServiceDatabaseList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: order
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: sort
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: universityName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取社会服务库列表
      tags:
      - SocialServiceDatabase
  /SSD/updateSocialServiceDatabase:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新社会服务库
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/EngineeringEducationDatabase.SocialServiceDatabase'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新社会服务库
      tags:
      - SocialServiceDatabase
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"